requirements.crawling.txt
# crawling
torch>=2.6.0
sentence-transformers>=2.7.0
langchain==0.3.25
qdrant-client==1.14.2
langchain_qdrant>=0.1.4,<0.2
langchain-huggingface>=0.1.0,<0.2
python-dotenv==1.1.0
google-generativeai>=0.3.0
python-dateutil>=2.8.2